some updates:
1. emulate ollama provider for VS code;
1. clean up warnings;
1. support Kimi-VL language model;
3. various code about vision.

+++ models/hunyuan.cpp
@@ -43,7 +43,7 @@ namespace dense
 
              terminate_ids.insert(end_of_text_token_id);
 
-            for (int i = 0; i < sizeof(extra_token_ids) / sizeof(extra_token_ids[0]); i++)
+            for (int i = 0; i < (int)(sizeof(extra_token_ids) / sizeof(extra_token_ids[0])); i++)
                 extra_token_ids[i] = SPECIAL_START_ID++;
 
             return size;
